yang BUGFIX overflow counter on enum statement

+++ src/parser_yang.h
@@ -42,14 +42,14 @@ struct lys_node_array{
     uint8_t unique;
     uint8_t tpdf;
     union {
-        uint8_t uni;
+        uint uni;
         uint16_t flags;
     };
     union {
-        uint8_t enm;
-        uint8_t pattern;
-        uint8_t bit;
-        uint8_t deviate;
+        uint enm;
+        uint pattern;
+        uint bit;
+        uint deviate;
     };
     uint16_t refine;
     uint16_t augment;
